Limit the graphical representation of deeply nested directories
This commit adds two features:

- Limit the graphical representation of deeply nested directories

The LEVELS_DEEP_THRESHOLD variable can be set to ignore nested files
and directories more than n levels in the graphical representation.

This is useful with large disk usage trees, where this limit can be beneficial
to reduce the tkdu memory usage.

- Limit the amount of error messages printed when parsing input

It's useful to see if there are parsing errors, but it is not often useful to
see too many errors.